+++ src/components/manage/Blocks/LeadImage/Edit.jsx
@@ -6,7 +6,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { compose } from 'redux';
-import { FormattedMessage, injectIntl } from 'react-intl';
+import { defineMessages, injectIntl } from 'react-intl';
 import cx from 'classnames';
 import { Message } from 'semantic-ui-react';
 
@@ -15,6 +15,13 @@ import { flattenToAppURL } from '@plone/volto/helpers';
 
 import imageBlockSVG from '@plone/volto/components/manage/Blocks/Image/block-image.svg';
 import { FormStateContext } from '@plone/volto/components/manage/Form/FormContext';
+
+const messages = defineMessages({
+  ImageBlockInputPlaceholder: {
+    id: "Upload a lead image in the 'Lead Image' content field.",
+    defaultMessage: "Upload a lead image in the 'Lead Image' content field.",
+  },
+});
 
 /**
  * Edit image block class.
@@ -62,6 +69,9 @@ class Edit extends Component {
   render() {
     const properties = this.context?.contextData?.formData || {};
     const { data } = this.props;
+    const placeholder =
+      this.props.data.placeholder ||
+      this.props.intl.formatMessage(messages.ImageBlockInputPlaceholder);
 
     return (
       <div
@@ -77,12 +87,7 @@ class Edit extends Component {
           <Message>
             <center>
               <img src={imageBlockSVG} alt="" />
-              <div className="message-text">
-                <FormattedMessage
-                  id="Upload a lead image in the 'Lead Image' content field."
-                  defaultMessage="Upload a lead image in the 'Lead Image' content field."
-                />
-              </div>
+              <div className="message-text">{placeholder}</div>
             </center>
           </Message>
         )}