+++ src/components/manage/Blocks/Listing/Edit.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage, injectIntl } from 'react-intl';
+import { defineMessages, injectIntl } from 'react-intl';
 
 import {
   SidebarPortal,
@@ -10,14 +10,18 @@ import {
 import { getBaseUrl } from '@plone/volto/helpers';
 import { useFormStateContext } from '@plone/volto/components/manage/Form/FormContext';
 
-const Edit = ({
-  data,
-  onChangeBlock,
-  block,
-  selected,
-  properties,
-  pathname,
-}) => {
+const messages = defineMessages({
+  results: {
+    id: 'Results preview',
+    defaultMessage: 'Results preview',
+  },
+  items: {
+    id: 'Contained items',
+    defaultMessage: 'Contained items',
+  },
+});
+
+const Edit = ({ data, onChangeBlock, block, selected, pathname, intl }) => {
   // componentDidMount
   React.useEffect(() => {
     if (!data.query) {
@@ -32,19 +36,15 @@ const Edit = ({
 
   const { contextData } = useFormStateContext();
   const { formData } = contextData;
+  const placeholder =
+    data.placeholder ||
+    (data?.query?.length
+      ? intl.formatMessage(messages.results)
+      : intl.formatMessage(messages.items));
 
   return (
     <>
-      {data?.query?.length === 0 && (
-        <FormattedMessage id="Contained items" defaultMessage="Contained items">
-          {(message) => <p className="items-preview">{message}</p>}
-        </FormattedMessage>
-      )}
-      {data?.query?.length > 0 && (
-        <FormattedMessage id="Results preview" defaultMessage="Results preview">
-          {(message) => <p className="items-preview">{message}</p>}
-        </FormattedMessage>
-      )}
+      <p className="items-preview">{placeholder}</p>
       <ListingBody
         data={data}
         properties={formData}