+++ src/components/manage/Blocks/Title/Edit.jsx
@@ -139,19 +139,26 @@ class Edit extends Component {
     if (__SERVER__) {
       return <div />;
     }
+
+    const placeholder =
+      this.props.data.placeholder ||
+      this.props.intl.formatMessage(messages.title);
+
     return (
       <Editor
         onChange={this.onChange}
         editorState={this.state.editorState}
         blockRenderMap={extendedBlockRenderMap}
         handleReturn={() => {
+          if (this.props.data.disableNewBlocks) {
+            return 'handled';
+          }
           this.props
             .onAddBlock(settings.defaultBlockType, this.props.index + 1)
             .then((id) => this.props.onSelectBlock(id));
-
           return 'handled';
         }}
-        placeholder={this.props.intl.formatMessage(messages.title)}
+        placeholder={placeholder}
         blockStyleFn={() => 'documentFirstHeading'}
         onUpArrow={() => {
           const selectionState = this.state.editorState.getSelection();