+++ src/components/manage/Blocks/ToC/Edit.jsx
@@ -14,10 +14,10 @@ import { useFormStateContext } from '@plone/volto/components/manage/Form/FormCon
  * @class Edit
  * @extends Component
  */
-const Edit = () => {
+const Edit = ({ data }) => {
   const { contextData } = useFormStateContext();
   const { formData } = contextData;
-  return <View properties={formData} />;
+  return <View properties={formData} data={data} />;
 };
 
 /**
@@ -27,6 +27,7 @@ const Edit = () => {
  */
 Edit.propTypes = {
   properties: PropTypes.objectOf(PropTypes.any).isRequired,
+  data: PropTypes.objectOf(PropTypes.any),
 };
 
 export default Edit;