+++ src/components/manage/Form/Form.jsx
@@ -3,8 +3,18 @@
  * @module components/manage/Form/Form
  */
 
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import { EditBlock, Field, Icon, Toast } from '@plone/volto/components';
+import {
+  blockHasValue,
+  difference,
+  FormValidation,
+  getBlocksFieldname,
+  getBlocksLayoutFieldname,
+  messages,
+} from '@plone/volto/helpers';
+import aheadSVG from '@plone/volto/icons/ahead.svg';
+import clearSVG from '@plone/volto/icons/clear.svg';
+import dragSVG from '@plone/volto/icons/drag.svg';
 import {
   findIndex,
   isEmpty,
@@ -13,76 +23,27 @@ import {
   mapValues,
   omit,
   pickBy,
-  uniq,
   without,
-  // isEqual,
 } from 'lodash';
 import move from 'lodash-move';
 import isBoolean from 'lodash/isBoolean';
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
+import { DragDropContext, Draggable, Droppable } from 'react-beautiful-dnd';
+import { injectIntl } from 'react-intl';
+import { Portal } from 'react-portal';
 import {
   Button,
   Container,
   Form as UiForm,
+  Message,
   Segment,
   Tab,
-  Message,
 } from 'semantic-ui-react';
-import { defineMessages, injectIntl } from 'react-intl';
 import { v4 as uuid } from 'uuid';
-import { Portal } from 'react-portal';
-
 import { toast } from 'react-toastify';
-import { EditBlock, Icon, Field, Toast } from '@plone/volto/components';
 import { settings } from '~/config';
-import dragSVG from '@plone/volto/icons/drag.svg';
 import { withFormStateContext } from '@plone/volto/components/manage/Form/FormContext';
-// import { FormStateContext, FormStateProvider } from './FormContext';
-
-import {
-  getBlocksFieldname,
-  getBlocksLayoutFieldname,
-  difference,
-  blockHasValue,
-} from '@plone/volto/helpers';
-
-import aheadSVG from '@plone/volto/icons/ahead.svg';
-import clearSVG from '@plone/volto/icons/clear.svg';
-import { DragDropContext, Draggable, Droppable } from 'react-beautiful-dnd';
-
-const messages = defineMessages({
-  addBlock: {
-    id: 'Add block…',
-    defaultMessage: 'Add block…',
-  },
-  required: {
-    id: 'Required input is missing.',
-    defaultMessage: 'Required input is missing.',
-  },
-  minLength: {
-    id: 'Minimum length is {len}.',
-    defaultMessage: 'Minimum length is {len}.',
-  },
-  uniqueItems: {
-    id: 'Items must be unique.',
-    defaultMessage: 'Items must be unique.',
-  },
-  save: {
-    id: 'Save',
-    defaultMessage: 'Save',
-  },
-  cancel: {
-    id: 'Cancel',
-    defaultMessage: 'Cancel',
-  },
-  error: {
-    id: 'Error',
-    defaultMessage: 'Error',
-  },
-  thereWereSomeErrors: {
-    id: 'There were some errors.',
-    defaultMessage: 'There were some errors.',
-  },
-});
 
 /**
  * Form container class.
@@ -115,6 +76,7 @@ export class Form extends Component {
     submitLabel: PropTypes.string,
     resetAfterSubmit: PropTypes.bool,
     isEditForm: PropTypes.bool,
+    isAdminForm: PropTypes.bool,
     title: PropTypes.string,
     error: PropTypes.shape({
       message: PropTypes.string,
@@ -124,6 +86,7 @@ export class Form extends Component {
     description: PropTypes.string,
     visual: PropTypes.bool,
     blocks: PropTypes.arrayOf(PropTypes.object),
+    requestError: PropTypes.string,
   };
 
   /**
@@ -138,6 +101,7 @@ export class Form extends Component {
     submitLabel: null,
     resetAfterSubmit: false,
     isEditForm: false,
+    isAdminForm: false,
     title: null,
     description: null,
     error: null,
@@ -147,6 +111,7 @@ export class Form extends Component {
     blocks: [],
     pathname: '',
     schema: {},
+    requestError: null,
   };
 
   /**
@@ -169,6 +134,9 @@ export class Form extends Component {
     this.onFocusPreviousBlock = this.onFocusPreviousBlock.bind(this);
     this.onFocusNextBlock = this.onFocusNextBlock.bind(this);
     this.handleKeyDown = this.handleKeyDown.bind(this);
+    this.onTabChange = this.onTabChange.bind(this);
+    this.onBlurField = this.onBlurField.bind(this);
+    this.onClickInput = this.onClickInput.bind(this);
 
     // We use these as instance fields, to be initialized in the render() meth
     // from the context provider
@@ -231,8 +199,79 @@ export class Form extends Component {
           : null,
       placeholderProps: {},
       isClient: false,
+      isFormPristine: true,
+      activeIndex: 0,
     };
     return state;
+  }
+
+  /**
+   * On updates caused by props change
+   * if errors from Backend come, these will be shown to their corresponding Fields
+   * also the first Tab to have any errors will be selected
+   * @param {Object} prevProps
+   */
+  componentDidUpdate(prevProps) {
+    if (this.props.formData?.['@id'] !== prevProps.formData?.['@id']) {
+      const newState = this.getInitialState(this.props);
+      this.setContextData(newState); // .then(() => this.setState(newState));;
+    }
+
+    let { requestError } = this.props;
+    let errors = {};
+    let activeIndex = 0;
+
+    if (requestError && prevProps.requestError !== requestError) {
+      errors = FormValidation.giveServerErrorsToCorrespondingFields(
+        requestError,
+      );
+      activeIndex = FormValidation.showFirstTabWithErrors({
+        errors,
+        schema: this.props.schema,
+      });
+
+      this.setContextData({
+        errors,
+        activeIndex,
+      });
+    }
+  }
+
+  /**
+   * Tab selection is done only by setting activeIndex in state
+   */
+  onTabChange(e, { activeIndex }) {
+    this.setContextData({ activeIndex });
+  }
+
+  /**
+   * If user clicks on input, the form will be not considered pristine
+   * this will avoid onBlur effects without interraction with the form
+   * @param {Object} e event
+   */
+  onClickInput(e) {
+    this.setContextData({ isFormPristine: false });
+  }
+
+  /**
+   * Validate fields on blur
+   * @method onBlurField
+   * @param {string} id Id of the field
+   * @param {*} value Value of the field
+   * @returns {undefined}
+   */
+  onBlurField(id, value) {
+    if (!this.contextData.isFormPristine) {
+      const errors = FormValidation.validateFieldsPerFieldset({
+        schema: this.props.schema,
+        formData: this.state.formData,
+        formatMessage: this.props.intl.formatMessage,
+      });
+
+      this.setContextData({
+        errors,
+      });
+    }
   }
 
   /**
@@ -248,49 +287,31 @@ export class Form extends Component {
   }
 
   /**
-   * Component did update
-   * @method componentDidUpdate
-   * @param {Object} prevProps Previous properties
-   * @returns {undefined}
-   */
-  componentDidUpdate(prevProps) {
-    if (this.props.formData?.['@id'] !== prevProps.formData?.['@id']) {
-      const newState = this.getInitialState(this.props);
-      this.setContextData(newState); // .then(() => this.setState(newState));;
-    }
-  }
-
-  /**
-   * Component will receive props
-   * @method componentWillReceiveProps
-   * @param {Object} nextProps Next properties
-   * @returns {undefined}
-   */
-  // UNSAFE_componentWillReceiveProps(nextProps) {
-  //   if (!isEqual(nextProps.formData, this.contextData.formData)) {
-  //     this.setContextData({ formData: nextProps.formData });
-  //   }
-  // }
-
-  /**
    * Change field handler
+   * Remove errors for changed field
    * @method onChangeField
    * @param {string} id Id of the field
    * @param {*} value Value of the field
    * @returns {undefined}
    */
   onChangeField(id, value) {
-    return this.setContextData({
-      formData: {
-        ...this.contextData.formData,
-        // We need to catch also when the value equals false this fixes #888
-        [id]: value || (value !== undefined && isBoolean(value)) ? value : null,
-      },
+    return this.setContextData((prevState) => {
+      const { errors } = prevState;
+      delete errors[id];
+      return {
+        errors,
+        formData: {
+          ...this.contextData.formData,
+          // We need to catch also when the value equals false this fixes #888
+          [id]:
+            value || (value !== undefined && isBoolean(value)) ? value : null,
+        },
+      };
     });
   }
 
   hideHandler = (data) => {
-    return !blockHasValue(data);
+    return !!data.fixed || !blockHasValue(data);
   };
 
   /**
@@ -321,7 +342,6 @@ export class Form extends Component {
    * @returns {undefined}
    */
   onMutateBlock(id, value) {
-    const idTrailingBlock = uuid();
     const blocksFieldname = getBlocksFieldname(this.contextData.formData);
     const blocksLayoutFieldname = getBlocksLayoutFieldname(
       this.contextData.formData,
@@ -329,6 +349,27 @@ export class Form extends Component {
     const index =
       this.contextData.formData[blocksLayoutFieldname].items.indexOf(id) + 1;
 
+    // Test if block at index is already a placeholder (trailing) block
+    const trailId = this.contextData.formData[blocksLayoutFieldname].items[
+      index
+    ];
+    if (trailId) {
+      const block = this.contextData.formData[blocksFieldname][trailId];
+      if (!blockHasValue(block)) {
+        this.setContextData({
+          formData: {
+            ...this.contextData.formData,
+            [blocksFieldname]: {
+              ...this.contextData.formData[blocksFieldname],
+              [id]: value || null,
+            },
+          },
+        });
+        return;
+      }
+    }
+
+    const idTrailingBlock = uuid();
     return this.setContextData({
       formData: {
         ...this.contextData.formData,
@@ -455,7 +496,7 @@ export class Form extends Component {
   }
 
   /**
-   * Submit handler
+   * Submit handler also validate form and collect errors
    * @method onSubmit
    * @param {Object} event Event object.
    * @returns {undefined}
@@ -464,48 +505,31 @@ export class Form extends Component {
     if (event) {
       event.preventDefault();
     }
-    const errors = {};
-    map(this.props.schema.fieldsets, (fieldset) =>
-      map(fieldset.fields, (fieldId) => {
-        const field = this.props.schema.properties[fieldId];
-        var data = this.contextData.formData[fieldId];
-        if (typeof data === 'string' || data instanceof String) {
-          data = data.trim();
-        }
-        if (this.props.schema.required.indexOf(fieldId) !== -1) {
-          if (field.type !== 'boolean' && !data) {
-            errors[fieldId] = errors[field] || [];
-            errors[fieldId].push(
-              this.props.intl.formatMessage(messages.required),
-            );
-          }
-          if (field.minLength && data && data.length < field.minLength) {
-            errors[fieldId] = errors[field] || [];
-            errors[fieldId].push(
-              this.props.intl.formatMessage(messages.minLength, {
-                len: field.minLength,
-              }),
-            );
-          }
-        }
-        if (field.uniqueItems && data && uniq(data).length !== data.length) {
-          errors[fieldId] = errors[field] || [];
-          errors[fieldId].push(
-            this.props.intl.formatMessage(messages.uniqueItems),
-          );
-        }
-      }),
-    );
+
+    const errors = FormValidation.validateFieldsPerFieldset({
+      schema: this.props.schema,
+      formData: this.contextData.formData,
+      formatMessage: this.props.intl.formatMessage,
+    });
+
     if (keys(errors).length > 0) {
-      return this.setContextData({
+      const activeIndex = FormValidation.showFirstTabWithErrors({
         errors,
-      }).then(() => {
-        Object.keys(errors).forEach((err) =>
-          toast.error(
-            <Toast error title={err} content={errors[err].join(', ')} />,
-          ),
-        );
+        schema: this.props.schema,
       });
+      this.setContextData(
+        {
+          errors,
+          activeIndex,
+        },
+        () => {
+          Object.keys(errors).forEach((err) =>
+            toast.error(
+              <Toast error title={err} content={errors[err].join(', ')} />,
+            ),
+          );
+        },
+      );
     } else {
       // Get only the values that have been modified (Edit forms), send all in case that
       // it's an add form
@@ -880,6 +904,7 @@ export class Form extends Component {
                               pathname={this.props.pathname}
                               block={block}
                               selected={contextData.selected === block}
+                              manage={this.props.isAdminForm}
                             />
                           </div>
                         </div>
@@ -897,7 +922,7 @@ export class Form extends Component {
                         width: `${placeholderProps.clientWidth}px`,
                         borderRadius: '3px',
                       }}
-                    />
+                    ></div>
                   )}
                 </div>
               )}
@@ -923,9 +948,11 @@ export class Form extends Component {
                             id={field}
                             formData={contextData.formData}
                             focus={false}
-                            value={contextData.formData[field]}
+                            value={contextData.formData?.[field]}
                             required={schema.required.indexOf(field) !== -1}
                             onChange={this.onChangeField}
+                            onBlur={this.onBlurField}
+                            onClick={this.onClickInput}
                             key={field}
                             error={contextData.errors[field]}
                           />
@@ -964,6 +991,8 @@ export class Form extends Component {
                     vertical: settings.verticalFormTabs,
                   }}
                   grid={{ paneWidth: 9, tabWidth: 3, stackable: true }}
+                  onTabChange={this.onTabChange}
+                  activeIndex={contextData.activeIndex}
                   panes={map(schema.fieldsets, (item) => ({
                     menuItem: item.title,
                     render: () => [
@@ -979,9 +1008,11 @@ export class Form extends Component {
                           formData={contextData.formData}
                           fieldSet={item.title.toLowerCase()}
                           focus={index === 0}
-                          value={contextData.formData[field]}
+                          value={contextData.formData?.[field]}
                           required={schema.required.indexOf(field) !== -1}
                           onChange={this.onChangeField}
+                          onBlur={this.onBlurField}
+                          onClick={this.onClickInput}
                           key={field}
                           error={contextData.errors[field]}
                         />
@@ -1026,6 +1057,8 @@ export class Form extends Component {
                     value={contextData.formData?.[field]}
                     required={schema.required.indexOf(field) !== -1}
                     onChange={this.onChangeField}
+                    onBlur={this.onBlurField}
+                    onClick={this.onClickInput}
                     key={field}
                     error={contextData.errors[field]}
                   />